requirements_vercel.txt
fastapi
uvicorn
sqlmodel
python-decouple
pywhatkit
qrcode
apscheduler
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-multipart 